requirements.txt
pygame
PyOpenGL
numpy
objloader-python